+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.3] Unreleased
+## [0.3.0] Unreleased
 ### Added
 - Ability to override command name via the [PoshBot.BotCommand] attribute.
 - Ability to set aliases for a command via the [PoshBot.BotCommand] attribute.